// es/config.d.ts
declare const CONFIG: {
    appId: string;
    prefix: string;
    mtrVer: string;
    stat_sw: boolean;
    stat_share_app: boolean;
    stat_pull_down_fresh: boolean;
    stat_page_scroll: boolean;
    stat_hide: boolean;
    stat_unload: boolean;
    stat_batch_send: boolean;
    stat_reach_bottom: boolean;
    stat_auto_click: boolean;
    stat_auto_expo: boolean;
    stat_auto_apperr: boolean;
    stat_location: boolean;
    stat_api: boolean;
    stat_api_success: boolean;
    stat_api_long_time: number;
    stat_api_fail: boolean;
    scene: {
        "1000": string;
        "1002": string;
        "1005": string;
        "1007": string;
        "1011": string;
        "1014": string;
        "1020": string;
        "1023": string;
        "1037": string;
        "1038": string;
        "1090": string;
        "1200": string;
        "1201": string;
        "1202": string;
        "1203": string;
        "1204": string;
        "1205": string;
        "1206": string;
        "1207": string;
        "1208": string;
        "1209": string;
        "1300": string;
        "0000": string;
    } | {
        '1000': string;
        '1001': string;
        '1005': string;
        '1006': string;
        '1007': string;
        '1008': string;
        '1010': string;
        '1011': string;
        '1012': string;
        '1013': string;
        '1014': string;
        '1017': string;
        '1019': string;
        '1020': string;
        '1022': string;
        '1023': string;
        '1024': string;
        '1025': string;
        '1026': string;
        '1027': string;
        '1028': string;
        '1029': string;
        '1030': string;
        '1031': string;
        '1032': string;
        '1034': string;
        '1035': string;
        '1036': string;
        '1037': string;
        '1038': string;
        '1039': string;
        '1042': string;
        '1043': string;
        '1044': string;
        '1045': string;
        '1046': string;
        '1047': string;
        '1048': string;
        '1049': string;
        '1052': string;
        '1053': string;
        '1054': string;
        '1056': string;
        '1057': string;
        '1058': string;
        '1059': string;
        '1064': string;
        '1067': string;
        '1068': string;
        '1069': string;
        '1071': string;
        '1072': string;
        '1073': string;
        '1074': string;
        '1077': string;
        '1078': string;
        '1079': string;
        '1081': string;
        '1082': string;
        '1084': string;
        '1088': string;
        '1089': string;
        '1090': string;
        '1091': string;
        '1092': string;
        '1095': string;
        '1096': string;
        '1097': string;
        '1099': string;
        '1102': string;
        '1103': string;
        '1104': string;
        '1106': string;
        '1107': string;
        '1113': string;
        '1114': string;
        '1124': string;
        '1125': string;
        '1126': string;
        '1129': string;
        '1131': string;
        '1133': string;
        '1135': string;
        '1146': string;
        '1148': string;
        '1150': string;
        '1153': string;
        '1169': string;
    };
    my: boolean;
    wx: boolean;
};
export default CONFIG;